==> 2017_07_compare_tracking/02_track_cp.Rmd <==
---
title: "load_data"
author: "TB"
date: "5/15/2017"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library("readr")
library("tibble")
library("dplyr")
library("tidyr")
library("stringr")
library("ggfortify")
library("cluster")
library("magrittr")
library("knitr")
#library("cytominer")
# load cytominer 

```

# Input 
Define the input folders and the csv files used to load the single cell data generated by cellprofiler.
```{r}
input_folder <- '../../../analysis/2017_morph_gradient_3/'
csv_in <- 'TrackeddetectedNeutros.csv'
```

# Output 
We use one output folder for the comparison and store two files
* tracks includes all files 
* clean_tracks stores the tracking information without bad frames identified by Gyde
```{r}
analysis_folder <- '../../../analysis/2017_07_compare_tracking/'

csv_tracks <- 'tracks_cp.csv'
csv_clean_tracks <- 'clean_tracks_cp.csv'
```

# Metadata 
```{r}
strata <- c('Metadata_condition','Metadata_dose','Metadata_id','Metadata_date','Metadata_matrix')
```

# Parameter
To recalculate the speed of the neutrophils in µm/minute we need the pixel size and the number of frames per minute.
```{r}
# scale speed from pixel/frame to µm/minute
pixel_size = 0.7819
frames_per_minute = 2
# frames are recorded each 30 seconds, i.e. 
speed_scale <- pixel_size * frames_per_minute

track_label = c('TrackObjects_Label_30')

```

# Load data 
The output of CellProfiler track module is loaded as CSV, tracked using cytominer::track and 
stored in the data folder as "track.csv". 
```{r}
# load csv files from all sub folders 
df <- data.frame()
for (iDir in list.dirs(input_folder)) {
  file_name <-  file.path(iDir, csv_in)
  if (file.exists(file_name)) {
    df <- rbind(df,suppressMessages(read_csv(file_name)))
  }
}

df %<>% 
  mutate(Metadata_timePoint = Metadata_frame) %>%
  dplyr::mutate(Metadata_timePoint = as.double(Metadata_timePoint)) %>%
  mutate(Metadata_frame = Metadata_timePoint) %>%
  select(Metadata_id, Metadata_date, Metadata_matrix, Metadata_condition, Metadata_dose, Metadata_frame, everything())
```


# track the neutros
```{r}
tracks <- df %>%
  group_by_(.dots =  c(strata,track_label)) %>%
  arrange(Metadata_timePoint) %>%
  cytominer::track(., c(strata,track_label)) %>%
  print()
```
## PostProcessing 
Calculate speed in µm/minute

```{r}
postprocessed <- tracks %>%
  mutate(Track_Speed = Track_Speed * speed_scale) %>%
  mutate(Track_Speed_max = Track_Speed_max * speed_scale) %>%
  mutate(Track_Speed_X = Track_Speed_X * speed_scale) %>%
  mutate(Track_Speed_Y = Track_Speed_Y * speed_scale) %>%
  mutate(Track_Distance_Traveled = pixel_size * Track_Distance_Traveled) %>%
  mutate(Track_Integrated_Distance_Traveled = pixel_size * Track_Integrated_Distance_Traveled) 
```

All na values are set to 'None' and all tracks are stored 
```{r}
postprocessed$Metadata_dose[is.na(postprocessed$Metadata_dose)] <- 'None'
write_csv(postprocessed, file.path(analysis_folder, csv_tracks))
```
# Cleanup 
Cleanup! Gyde provided a list with frames / videos with an possible technical 
error. We remove these data.
Gydes list:

09.07.2013 GN0039: Kontrolle - frame 121, 

  * 0,5µM - frame 1 und 2, 
  * 50µM -frame 1-7, 
  * Vehikel - frame 1
  
11.06.2013 GN0033: 

  * 5µm - frame 1-45! Wenn diese frames nicht exkludiert werden können, muss der Tag exkludiert werden
  * 11.07.2013 GN0040: 
  * Kontrolle - frame 1-5, 
  * 0,5µM - 1-7, 
  * 50µm - 1-9

12.07.2013 GN0041: 

  * Kontrolle - 1-10, 
  * 0,5µm- 1-10, 
  * 5µm - 1-10, 
  * 50µm - 1-10

25.04.2013 GN0017: 

  * 0,5µm - frame 1-50!, 
  * 50µm- 1-25!, 
  * Vehikel - 1-25! Wenn diese frames nicht exkludiert werden können, muss der Tag exkludiert werden

15.07.2013 BD0053: 

  * 0,5µm - 1-10, 
  * 5µm - 1-5, 
  * 50µm - 1-5, 
  * Vehikel - 1-5

17.07.2013 BD0054: 

  * Kontrolle - 1-5, 
  * 0,5µm - 1-5, 
  * Vehikel - 1-10

19.07.2013 BD0055: 

  * Kontrolle - 1-10, 
  * 0,5µm - 1-10, 
  * 5µm - 1-5, 
  * 50µm - 1-5

17.09.2013 EJ0005: 

  * 0,5µm - 1-10, 
  * Vehikel HEM - 1, 
  * Vehikel FN - 1-5

18.09.2013 BD0069: 

  * 5µm - 1-5, 
  * 50µm - 1-7, 
  * Vehikel FN - 1-10

19.09.2013 EJ0006: 

  * Vehikel FN - 115-121

20.09.2013 BD0071: 

  * Kontrolle - 1-5, 
  * 0,5µm - 1-5, 
  * Vehikel HEM - 1-10,
  * Vehikel FN - 1-5

26.09.2013 GN0050: 

  * Kontrolle - 1-10, 
  * 50µm - 1-5,   
  * Vehikel HEM - 1-5,
  * Vehikel FN - 1-10

14.11.2013 GN0055: 

  * Kontrolle - 1-25, 
  * 5µm - 1-5, 
  * Vehikel HEM - 1-5

27.09.2013 BD0073: 

  * Kontrolle - 1-5, 
  * 5µm - 1-10, 
  * 50µm - 1-7, 
  * Vehikel HEM- 1-5, 
  * Vehikel FN 1-25! Wenn diese frames nicht exkludiert werden können, hier kein FN vs HEM Vergleich!

30.09.2013 GN0051: 

  * Kontrolle - 1-10, 
  * 5µm - 1-10, 
  * Vehikel HEM - 1-5,
  * Vehikel FN - 1-70! Wenn diese frames nicht exkludiert werden können, hier kein FN vs HEM Vergleich!

03.10.2013 GN0052: 

  * Kontrolle - 1-10, 
  * 5µm - 1-5, 
  * Vehikel HEM - 1-5


Same list in R: 
```{r}

clean_df <- df %>% 
  filter( !((Metadata_id == 'gn0039') & (Metadata_condition == 'control') &  (Metadata_frame == 121) ) ) %>% 
  filter( !((Metadata_id == 'gn0039') & (Metadata_condition == 'vehicle') &  (Metadata_frame == 1) ) ) %>% 
  filter( !((Metadata_id == 'gn0039') & (Metadata_dose == 0.5) &  (Metadata_frame < 3) ) ) %>%
  filter( !((Metadata_id == 'gn0039') & (Metadata_dose == 50) &  (Metadata_frame < 8) ) ) %>% 
  
  filter( !((Metadata_id == 'gn0033') & (Metadata_dose == 5) &  (Metadata_frame < 46) ) ) %>%
  
  filter( !((Metadata_id == 'gn0040') & (Metadata_condition == 'control') & (Metadata_frame < 6) ) ) %>%
  filter( !((Metadata_id == 'gn0040') & (Metadata_dose == 0.5) &  (Metadata_frame < 8) ) ) %>%
  filter( !((Metadata_id == 'gn0040') & (Metadata_dose == 50) &  (Metadata_frame < 10) ) ) %>%
  
  filter( !((Metadata_id == 'gn0041') & (Metadata_condition == 'control') & (Metadata_frame < 11) ) ) %>%
  filter( !((Metadata_id == 'gn0041') & (Metadata_dose == 0.5) &  (Metadata_frame < 11) ) ) %>%
  filter( !((Metadata_id == 'gn0041') & (Metadata_dose == 5) &  (Metadata_frame < 11) ) ) %>%
  filter( !((Metadata_id == 'gn0041') & (Metadata_dose == 50) &  (Metadata_frame < 11) ) ) %>%
  
  filter( !((Metadata_id == 'gn0017') & (Metadata_condition == 'vehicle') & (Metadata_frame < 25) ) ) %>%
  filter( !((Metadata_id == 'gn0017') & (Metadata_dose == 0.5) &  (Metadata_frame < 51) ) ) %>%
  filter( !((Metadata_id == 'gn0017') & (Metadata_dose == 50) &  (Metadata_frame < 26) ) ) %>%
  
  filter( !((Metadata_id == 'gn0017') & (Metadata_condition == 'vehicle') & (Metadata_frame < 25) ) ) %>%
  filter( !((Metadata_id == 'gn0017') & (Metadata_dose == 0.5) &  (Metadata_frame < 51) ) ) %>%
  filter( !((Metadata_id == 'gn0017') & (Metadata_dose == 50) &  (Metadata_frame < 26) ) ) %>%
  
  filter( !((Metadata_id == 'bd0053') & (Metadata_condition == 'vehicle') & (Metadata_frame < 6) ) ) %>%
  filter( !((Metadata_id == 'bd0053') & (Metadata_dose == 0.5) &  (Metadata_frame < 11) ) ) %>%
  filter( !((Metadata_id == 'bd0053') & (Metadata_dose == 5) &  (Metadata_frame < 6) ) ) %>%
  filter( !((Metadata_id == 'bd0053') & (Metadata_dose == 50) &  (Metadata_frame < 6) ) ) %>%
  
  filter( !((Metadata_id == 'bd0054') & (Metadata_condition == 'control') & (Metadata_frame < 6) ) ) %>%
  filter( !((Metadata_id == 'bd0054') & (Metadata_condition == 'vehicle') & (Metadata_frame < 11) ) ) %>%
  filter( !((Metadata_id == 'bd0054') & (Metadata_dose == 0.5) &  (Metadata_frame < 6) ) ) %>%
  
  filter( !((Metadata_id == 'bd0054') & (Metadata_condition == 'control') & (Metadata_frame < 6) ) ) %>%
  filter( !((Metadata_id == 'bd0054') & (Metadata_condition == 'vehicle') & (Metadata_frame < 11) ) ) %>%
  filter( !((Metadata_id == 'bd0054') & (Metadata_dose == 0.5) &  (Metadata_frame < 6) ) ) %>%
  
  filter( !((Metadata_id == 'bd0055') & (Metadata_condition == 'control') & (Metadata_frame < 11) ) ) %>%
  filter( !((Metadata_id == 'bd0055') & (Metadata_dose == 0.5) &  (Metadata_frame < 11) ) ) %>%
  filter( !((Metadata_id == 'bd0055') & (Metadata_dose == 5) &  (Metadata_frame < 6) ) ) %>%
  filter( !((Metadata_id == 'bd0055') & (Metadata_dose == 50) &  (Metadata_frame < 6) ) ) %>%
  
  filter( !((Metadata_id == 'bd0055') & (Metadata_condition == 'control') & (Metadata_frame < 11) ) ) %>%
  filter( !((Metadata_id == 'bd0055') & (Metadata_dose == 0.5) &  (Metadata_frame < 11) ) ) %>%
  filter( !((Metadata_id == 'bd0055') & (Metadata_dose == 5) &  (Metadata_frame < 6) ) ) %>%
  filter( !((Metadata_id == 'bd0055') & (Metadata_dose == 50) &  (Metadata_frame < 6) ) ) %>%
  
  filter( !((Metadata_id == 'ej0005') & (Metadata_condition == 'vehicle') & (Metadata_matrix == 'HEM') & (Metadata_frame < 2) ) ) %>%
  filter( !((Metadata_id == 'ej0005') & (Metadata_condition == 'vehicle') & (Metadata_matrix == 'FN') & (Metadata_frame < 6) ) ) %>%
  filter( !((Metadata_id == 'ej0005') & (Metadata_dose == 0.5) &  (Metadata_frame < 11) ) ) %>%
  
  filter( !((Metadata_id == 'bd0069') & (Metadata_condition == 'vehicle') & (Metadata_matrix == 'FN') & (Metadata_frame < 11) ) ) %>%
  filter( !((Metadata_id == 'bd0069') & (Metadata_dose == 5) &  (Metadata_frame < 6) ) ) %>%
  filter( !((Metadata_id == 'bd0069') & (Metadata_dose == 50) &  (Metadata_frame < 8) ) ) %>%
  
  filter( !((Metadata_id == 'ej0006') & (Metadata_condition == 'vehicle') & (Metadata_matrix == 'FN') & (Metadata_frame > 114) ) ) %>%
  
  filter( !((Metadata_id == 'bd0071') & (Metadata_condition == 'control') & (Metadata_frame < 6) ) ) %>%
  filter( !((Metadata_id == 'bd0071') & (Metadata_condition == 'vehicle') & (Metadata_matrix == 'HEM') & (Metadata_frame < 11) ) ) %>%
  filter( !((Metadata_id == 'bd0071') & (Metadata_condition == 'vehicle') & (Metadata_matrix == 'FN') & (Metadata_frame < 6) ) ) %>%
  filter( !((Metadata_id == 'bd0071') & (Metadata_dose == 0.5) &  (Metadata_frame < 6) ) ) %>%
  
  filter( !((Metadata_id == 'gn0050') & (Metadata_condition == 'control') & (Metadata_frame < 11) ) ) %>%
  filter( !((Metadata_id == 'gn0050') & (Metadata_condition == 'vehicle') & (Metadata_matrix == 'HEM') & (Metadata_frame < 6) ) ) %>%
  filter( !((Metadata_id == 'gn0050') & (Metadata_condition == 'vehicle') & (Metadata_matrix == 'FN') & (Metadata_frame < 11) ) ) %>%
  filter( !((Metadata_id == 'gn0050') & (Metadata_dose == 50) &  (Metadata_frame < 6) ) ) %>%
  
  filter( !((Metadata_id == 'gn0055') & (Metadata_condition == 'control') & (Metadata_frame < 26) ) ) %>%
  filter( !((Metadata_id == 'gn0055') & (Metadata_condition == 'vehicle') & (Metadata_matrix == 'HEM') & (Metadata_frame < 6) ) ) %>%
  filter( !((Metadata_id == 'gn0055') & (Metadata_dose == 5) &  (Metadata_frame < 6) ) ) %>%
  
  filter( !((Metadata_id == 'bd0073') & (Metadata_condition == 'control') & (Metadata_frame < 6) ) ) %>%
  filter( !((Metadata_id == 'bd0073') & (Metadata_condition == 'vehicle') & (Metadata_matrix == 'HEM') & (Metadata_frame < 6) ) ) %>%
  filter( !((Metadata_id == 'bd0073') & (Metadata_condition == 'vehicle') & (Metadata_matrix == 'FN') & (Metadata_frame < 26) ) ) %>%
  filter( !((Metadata_id == 'bd0073') & (Metadata_dose == 5) &  (Metadata_frame < 11) ) ) %>%
  filter( !((Metadata_id == 'bd0073') & (Metadata_dose == 50) &  (Metadata_frame < 8) ) ) %>%
  
  filter( !((Metadata_id == 'gn0051') & (Metadata_condition == 'control') & (Metadata_frame < 11) ) ) %>%
  filter( !((Metadata_id == 'gn0051') & (Metadata_condition == 'vehicle') & (Metadata_matrix == 'HEM') & (Metadata_frame < 6) ) ) %>%
  filter( !((Metadata_id == 'gn0051') & (Metadata_condition == 'vehicle') & (Metadata_matrix == 'FN') & (Metadata_frame < 71) ) ) %>%
  filter( !((Metadata_id == 'gn0051') & (Metadata_dose == 5) &  (Metadata_frame < 11) ) ) %>%
  
  filter( !((Metadata_id == 'gn0052') & (Metadata_condition == 'control') & (Metadata_frame < 11) ) ) %>%
  filter( !((Metadata_id == 'gn0052') & (Metadata_condition == 'vehicle') & (Metadata_matrix == 'HEM') & (Metadata_frame < 6) ) ) %>%
  filter( !((Metadata_id == 'gn0052') & (Metadata_dose == 5) &  (Metadata_frame < 6) ) ) %>%
  
  print
```

Track the "cleaned" data
```{r}
clean_tracks <- clean_df %>%
  group_by_(.dots =  c(strata,track_label)) %>%
  arrange(Metadata_timePoint) %>%
  cytominer::track(., c(strata,track_label)) %>%
  group_by_(.dots = strata) %>% 
  print()
```


## PostProcessing 
Calculate speed in µm/minute
```{r}
clean_postprocessed <- clean_tracks %>%
  mutate(Track_Speed = Track_Speed * speed_scale) %>%
  mutate(Track_Speed_max = Track_Speed_max * speed_scale) %>%
  mutate(Track_Speed_X = Track_Speed_X * speed_scale) %>%
  mutate(Track_Speed_Y = Track_Speed_Y * speed_scale) %>%
  mutate(Track_Distance_Traveled = pixel_size * Track_Distance_Traveled) %>%
  mutate(Track_Integrated_Distance_Traveled = pixel_size * Track_Integrated_Distance_Traveled) 
```


All na values are set to 'None' and all tracks are stored 
```{r}
clean_postprocessed$Metadata_dose[is.na(clean_postprocessed$Metadata_dose)] <- 'None'
write_csv(clean_postprocessed, file.path(analysis_folder, csv_clean_tracks))
```
